Wrap GRPC errors to include stack trace.
Initial issue I was trying to solve is that in case of unknown error
trail.FromGRPC rebuilds the error object from scratch. It leads to a
problem that one cannot reason anymore what the original error it was.
For example one would want to check trace.IsEOF(err) on the result of
stream.Recv() but io.EOF is not a GRPC error, so FromGRPC treats it
as unknown.

So in case of unknown error it's better to just wrap the original error.

But it'd be weird if some errors will be returned wrapped and some
unwrapped. I think that trail.FromGRPC should return every error wrapped
so stack trace will be included too which is the initial goal of the
trace library.

Also I had to introduce some weird internal methods in order to have a control
on how library captures the stack: trace caller should be on top of the stack
without any frames from files like trace.go, errors.go, trail.go, etc.