Remove mutating state methods - breaking change (#85)
PR https://github.com/gravitational/trace/pull/83 changed the API behavior. Currently, `AddUserMessage()` and similar do not change the current struct but return a modified copy. This makes errors like that:

```go
err.AddUserMessage("extra args 1")
err.AddUserMessage("extar args 2")
// err doesn't contain any user messages, but it used to.
```

This PR converts `Add.*` methods to `With*.` functions to make it clear that the returned error should be used and the state it not mutated. This should also eliminate the race condition mentioned in the previous PR.