feat!: Make "IsError" methods work with Go-wrapped errors (#95)
"IsError" methods, like "IsBadParameter" or "IsNotFound", fail to identify the
corresponding errors if Go wrapping is used in the chain. This PR fixes that,
allowing `trace.Wrap` and `fmt.Errorf` to coexist gracefully.

Comparisons are now based on `errors.As` (or equivalent logic), not the previous
interface-casting mechanisms.

I've also taken the opportunity to address some other issues within the
codebase.